task: fix ssl issues
Note that this is not the only, or even the best way to fix the ssl
issues, but this at least works. In order to make use of vite generated
ssl certs, I think you first need to run

pnpx vite --host 0.0.0.0

which will just use the vite config and will generate the certs at

/home/vscode/.vite-plugin-mkcert/*.pem

This doesn't actually run the server we want to serve to you then stop
the vite server and instead run

pnpm run dev. I hope this works for you 😊